Remove params
