+++ sample_service/lib/sample_service.rb
@@ -7,7 +7,6 @@ require_relative 'sample_service/builders'
 require_relative 'sample_service/client'
 require_relative 'sample_service/errors'
 require_relative 'sample_service/paginators'
-require_relative 'sample_service/params'
 require_relative 'sample_service/parsers'
 require_relative 'sample_service/stubs'
 require_relative 'sample_service/types'
