+++ sample_service/lib/sample_service/types.rb
@@ -1,13 +1,44 @@
 module SampleService
   module Types
 
-    StructuredEvent = Struct.new(:message, keyword_init: true)
+    StructuredEvent = Struct.new(
+      :message,
+      keyword_init: true
+    ) do
+      def self.build(params)
+        type = new
+        type.message = params[:message]
+        type
+      end
+    end
 
     class EventStream < SimpleDelegator
       class Start < EventStream; end
       class End < EventStream; end
       class Log < EventStream; end
       class Unknown < EventStream; end
+
+      def self.build(params)
+        return params if params.is_a?(EventStream)
+
+        if params.compact.size > 1
+          raise ArgumentError,
+                "EventStream must have exactly one member, got: #{params}"
+        end
+        key, value = params.flatten
+        case key
+        when :start
+          Start.new(
+            StructuredEvent.build(params[:start])
+          )
+        when :end
+          End.new(
+            StructuredEvent.build(params[:end])
+          )
+        when :log
+          Log.new(value)
+        end
+      end
     end
 
     # Permitted params for a High Score
@@ -24,7 +55,14 @@ module SampleService
       :game,
       :score,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.game = params[:game]
+        type.score = params[:score]
+        type
+      end
+    end
 
     # Modeled attributes for a High Score
     #
@@ -55,7 +93,17 @@ module SampleService
       :created_at,
       :updated_at,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.id = params[:id]
+        type.game = params[:game]
+        type.score = params[:score]
+        type.created_at = params[:created_at]
+        type.updated_at = params[:updated_at]
+        type
+      end
+    end
 
     # Input structure for GetHighScore
     #
@@ -66,7 +114,13 @@ module SampleService
     GetHighScoreInput = Struct.new(
       :id,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.id = params[:id]
+        type
+      end
+    end
 
     # Output structure for GetHighScore
     #
@@ -77,7 +131,13 @@ module SampleService
     GetHighScoreOutput = Struct.new(
       :high_score,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.high_score = HighScoreAttributes.new(params[:high_score]) if params[:high_score]
+        type
+      end
+    end
 
     # Input structure for CreateHighScore
     #
@@ -88,7 +148,13 @@ module SampleService
     CreateHighScoreInput = Struct.new(
       :high_score,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.high_score = HighScoreParams.build(params[:high_score]) if params[:high_score]
+        type
+      end
+    end
 
     # Output structure for CreateHighScore
     #
@@ -104,7 +170,14 @@ module SampleService
       :high_score,
       :location,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.high_score = HighScoreAttributes.new(params[:high_score]) if params[:high_score]
+        type.location = params[:location]
+        type
+      end
+    end
 
     # Input structure for UpdateHighScore
     #
@@ -120,7 +193,14 @@ module SampleService
       :id,
       :high_score,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.id = params[:id]
+        type.high_score = HighScoreParams.build(params[:high_score]) if params[:high_score]
+        type
+      end
+    end
 
     # Output structure for UpdateHighScore
     #
@@ -131,7 +211,13 @@ module SampleService
     UpdateHighScoreOutput = Struct.new(
       :high_score,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.high_score = HighScoreAttributes.new(params[:high_score]) if params[:high_score]
+        type
+      end
+    end
 
     # Input structure for DeleteHighScore
     #
@@ -142,20 +228,38 @@ module SampleService
     DeleteHighScoreInput = Struct.new(
       :id,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.id = params[:id]
+        type
+      end
+    end
 
     # Output structure for DeleteHighScore
     DeleteHighScoreOutput = Struct.new(
       nil,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type
+      end
+    end
 
     # Input structure for ListHighScores
     ListHighScoresInput = Struct.new(
       :max_results,
       :next_token,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.next_token = params[:next_token]
+        type.max_results = params[:max_results]
+        type
+      end
+    end
 
     # Output structure for ListHighScores
     #
@@ -167,19 +271,41 @@ module SampleService
       :next_token,
       :high_scores,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.next_token = params[:next_token]
+        type.high_scores = params[:high_scores].map do |high_score|
+          HighScoreAttributes.build(high_score)
+        end
+        type
+      end
+    end
 
     # Input and Output structure for Stream
     StreamInputOutput = Struct.new(
       :stream_id,
       :blob
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.stream_id = params[:stream_id]
+        type.blob = params[:blob]
+        type
+      end
+    end
 
     # Raised when high score is invalid
     UnprocessableEntityError = Struct.new(
       :errors,
       keyword_init: true
-    )
+    ) do
+      def self.build(params)
+        type = new
+        type.errors = Hash[params[:errors].map { |k, v| [k, v] }]
+        type
+      end
+    end
 
   end
 end
